+++ package.json
@@ -94,6 +94,7 @@
     "mime-types": "^2.1.14",
     "mkdirp": "^3.0.1",
     "query-string": "^7.1.1",
+    "stream-json": "^1.7.5",
     "through2": "^4.0.2",
     "web-encoding": "^1.1.5",
     "xml": "^1.0.0",
