+++ src/extensions.ts
@@ -16,8 +16,9 @@
 
 import * as stream from 'node:stream'
 
+import { isBoolean, isNumber, isString } from './assert.ts'
 import * as errors from './errors.ts'
-import { isBoolean, isNumber, isString, isValidBucketName, isValidPrefix, pipesetup, uriEscape } from './helpers.ts'
+import { isValidBucketName, isValidPrefix, pipesetup, uriEscape } from './helpers.ts'
 import * as transformers from './transformers.ts'
 import type { Client } from './typed-client2.ts'
