+++ src/s3-endpoints.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { isString } from './helpers.ts'
+import { isString } from './assert.ts'
 
 // List of currently supported endpoints.
 const awsS3Endpoint = {
