+++ src/transformers.ts
@@ -20,8 +20,8 @@ import type * as stream from 'node:stream'
 
 import Through2 from 'through2'
 
+import { isFunction } from './assert.ts'
 import * as errors from './errors.ts'
-import { isFunction } from './helpers.ts'
 import * as xmlParsers from './xml-parsers.ts'
 
 // getConcater returns a stream that concatenates the input and emits
